refactor: api 요청 이름에 맞게 타입 네이밍 수정
