--- src/types/receiver.d.ts
@@ -1,16 +1,0 @@
-export type Receiver = {
-  name: string;
-  phone: string;
-  count: number;
-};
-
-export type FormValues = {
-  modalReceiver: Receiver[];
-  receiver: Receiver[];
-};
-
-export type OrderFormValue = {
-  message: string;
-  sender: string;
-  receiver: Receiver[];
-};
